# MarketPlaceWebServiceExpress/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.0.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47300/service/v1
    description: Local development
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
...
    ImageMeta:
      type: object
      properties:
        image_id:
          type: string
          description: Same as item_code
        guid_code:
          type: string
          format: uuid
...
paths:
...
  /getImageList:
    get:
      tags: [Images]
      summary: รายการ image metadata ของสินค้า
      description: ดึงจาก demo_images database — เรียงตาม image_order ASC
      security: []
      parameters:
        - { name: item_code, in: query, required: true, schema: { type: string }, example: "ITM001" }
      responses:
        '200':
          description: รายการ guid ของรูปภาพ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ImageMeta' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /images:
    get:
      tags: [Images]
      summary: ดาวน์โหลดภาพตาม item_code
      description: |
        ส่งคืน binary PNG ของรูปแรกของสินค้า

        รูปสินค้าถูก cache 1 ชั่วโมง หลัง cache หมดอายุ request ถัดไปจะกลับไปดึงรูปจาก server ใหม่
        Cache-Control: `public, max-age=3600, must-revalidate`
      security: []
      parameters:
        - { name: item_code,      in: query,  required: true, schema: { type: string },         example: "ITM001" }
      responses:
        '200':
          description: PNG image binary
          headers:
            Cache-Control:
              schema: { type: string }
              example: "public, max-age=3600, must-revalidate"
            Content-Length:
              schema: { type: integer }
          content:
            image/png:
              schema: { type: string, format: binary }
        '400':
          description: item_code ไม่ได้ส่งมา
          content:
            text/plain:
              schema: { type: string }
              example: "ERROR: item_code is required"
        '404':
          description: ไม่พบรูปภาพ
          content:
            text/plain:
              schema: { type: string }
              example: "ERROR: image not found"
  /imagesguid:
    get:
      tags: [Images]
      summary: ดาวน์โหลดภาพตาม guid_code
      description: |
        ใช้สำหรับรูปที่ 2, 3... ของสินค้า (ได้ guid จาก `/getImageList`)

        **ต่างจาก /images**: ถ้าไม่พบ guid → HTTP 200 + **empty buffer (0 bytes)** ไม่ใช่ 404

        ตรวจ `Content-Length: 0` ใน client เพื่อรู้ว่ารูปไม่มี
      security: []
      parameters:
        - { name: guid_code,      in: query,  required: true, schema: { type: string, format: uuid }, example: "550e8400-e29b-41d4-a716-446655440000" }
      responses:
        '200':
          description: |
            PNG image binary — หรือ empty buffer (Content-Length=0) ถ้าไม่พบ
          headers:
            Cache-Control:
              schema: { type: string }
              example: "public, max-age=3600, must-revalidate"
            Content-Length:
              schema: { type: integer }
              description: 0 ถ้าไม่พบรูป
          content:
            image/png:
              schema: { type: string, format: binary }

  # ─── POS ─────────────────────────────────────────────────────────────────────
...